Task_44 Refactoring
