--- app/src/main/java/com/example/moviesearch/data/ApiConstants.kt
@@ -1,6 +1,0 @@
-package com.example.moviesearch.data
-
-object ApiConstants {
-    const val BASE_URL = "https://api.themoviedb.org/"
-    const val IMAGES_URL = "https://image.tmdb.org/t/p/"
-}
