--- app/src/main/java/com/example/moviesearch/data/Entity/TmdbResults.kt
@@ -1,14 +1,0 @@
-package com.example.moviesearch.data.Entity
-
-import com.google.gson.annotations.SerializedName
-
-data class TmdbResults(
-    @SerializedName("page")
-    val page: Int,
-    @SerializedName("results")
-    val tmdbFilms: List<TmdbFilm>,
-    @SerializedName("total_pages")
-    val totalPages: Int,
-    @SerializedName("total_results")
-    val totalResults: Int
-)
