+++ remote_module/src/main/java/com/example/remote_module/RemoteModule.kt
@@ -1,4 +1,46 @@
 package com.example.remote_module
 
+import com.example.remote_module.entity.ApiConstants
+import dagger.Module
+import dagger.Provides
+import hu.akarnokd.rxjava3.retrofit.RxJava3CallAdapterFactory
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import java.util.concurrent.TimeUnit
+import javax.inject.Singleton
+
+@Module
 class RemoteModule {
+    @Provides
+    @Singleton
+    fun provideOkHttpClient(): OkHttpClient = OkHttpClient.Builder()
+        //Настраиваем таймауты для медленного интернета
+        .callTimeout(30, TimeUnit.SECONDS)
+        .readTimeout(30, TimeUnit.SECONDS)
+        //Добавляем логгер
+        .addInterceptor(HttpLoggingInterceptor().apply {
+            if (BuildConfig.DEBUG) {
+                level = HttpLoggingInterceptor.Level.BASIC
+            }
+        })
+        .build()
+
+    @Provides
+    @Singleton
+    fun provideRetrofit(okHttpClient: OkHttpClient): Retrofit = Retrofit.Builder()
+        //Указываем базовый URL из констант
+        .baseUrl(ApiConstants.BASE_URL)
+        //Добавляем конвертер
+        .addConverterFactory(GsonConverterFactory.create())
+        //Добавляем поддержку RxJava
+        .addCallAdapterFactory(RxJava3CallAdapterFactory.create())
+        //Добавляем кастомный клиент
+        .client(okHttpClient)
+        .build()
+
+    @Provides
+    @Singleton
+    fun provideTmdbApi(retrofit: Retrofit): TmdbApi = retrofit.create(TmdbApi::class.java)
 }
